backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Meir <maornet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
